Refactor extension-related types and functions for improved type safety and clarity
