+++ src/components/ButtonGroupWithDropdown/ButtonGroupWithDropdown.tsx
@@ -14,7 +14,7 @@ interface ButtonGroupWithDropdownProps extends BoxProps {
   options: Array<{
     label: string;
     testId?: string;
-    onSelect: (...args: any[]) => void;
+    onSelect: <T extends object>(params: T) => void;
   }>;
   testId?: string;
   disabled?: boolean;
